restructure test folders to match library namespace
